Playstore/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.9
django-allauth==0.60.0
django-environ==0.12.0
django-phonenumber-field==7.3.0
django-twilio==0.14.3.2
idna==3.6
oauthlib==3.2.2
package_name==0.1
phonenumbers==8.13.27
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
razorpay==1.4.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
sqlparse==0.4.4
twilio==7.17.0
tzdata==2023.4
urllib3==2.1.0
